requirements.txt
aws-cdk-lib==2.142.1
aws-lambda-powertools==2.29.0
cdk-nag==2.27.176
constructs
stringcase==1.2.0
PyYAML
boto3==1.34.9
cdk-aws-iotfleetwise==0.3.2
aws-cdk.aws-apigatewayv2-integrations-alpha